requirements.txt
streamlit
pdfplumber
pytesseract
pandas
scikit-learn
nltk
Pillow